compile.sh: check DMA and scheduler for warnings
The scheduler is too important to ignore that affect the MM a lot
especially for performance reasons. The DMA is kind of similar to the MM
and IOMMU. The focus here is not to worry too much about those
driver-specific things like DMA engine on devices and RDMA if possible.

Signed-off-by: Qian Cai <cai@lca.pw>