config: select CONFIG_PROVE_LOCKING=y for all
Historically, there was a lockdep false positive with XFS [1] that was
easy to trigger. Once it is triggered, the lockdep will disable itself,
so there is no point to enable it without silencing it first. Over time,
the XFS developers are gearing towards fixing those. The latest update
[2] certainly look encouraging.

Recently, there is a system hang issue [3] that has proved lockdep is
still valuable for debugging, enable it for all architectures, but keep
in mind that there are going to be a lot of false positives need to deal
with, especially from debugging code paths, that very few people are
actively testing.

[1] https://lore.kernel.org/lkml/20190109205329.2486-1-cai@lca.pw/
[2] https://lore.kernel.org/lkml/20190822131739.GB1119@dread.disaster.area/
[3] https://lore.kernel.org/lkml/20190911011008.GA4420@jagdpanzerIV/

Signed-off-by: Qian Cai <cai@lca.pw>