s390: enable KVM
Tested it on z13 z/VM 6.4.0 though not expecting to run this kernel as a
KVM guest yet. There is a issue to run with "-serial mon:stdio" due to
"-nographic", but it is fine without.

qemu-kvm: cannot use stdio by multiple character devices
qemu-kvm: could not connect sclp console to character backend 'stdio'

Signed-off-by: Qian Cai <cai@lca.pw>