kvm.sh: enable SR-IOV and PCI passthrough
It takes a parameter like "0000:24:00.3" as a passthrough device number.
Intel currently gives an error for PCI passthrough,

"DMAR: Device is ineligible for IOMMU domain attach due to platform RMRR
requirement.  Contact your platform vendor."

or it could trigger a boot crash with KCSAN and the hpsa driver.
Eventually, it should be possible to run SR-IOV on Intel.

Signed-off-by: Qian Cai <cai@lca.pw>